projectLists finalized
