+++ index.html
@@ -6,11 +6,10 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <link rel="stylesheet" href="resources/css/style.css">
   <link href="https://fonts.googleapis.com/css2?family=Manrope:wght@300;500;700&display=swap" rel="stylesheet">
-  <script src="resources/js/projects.js"></script>
   <title>Dashboard</title>
 </head>
 
-<body class="flex-box" onload="getProjects()">
+<body class="flex-box">
   <!--    Side panel section-->
   <section class="side-panel">
 
@@ -216,8 +215,9 @@
     <!--      Footer ends here-->
   </section>
 
-
-  <script src="resources/js/common.js"></script>
+  
+  <script type="module" src="resources/js/api.js"></script>
+  <script type="module" src="resources/js/common.js"></script>
 </body>
 
 </html>
