+++ resources/css/style.css
@@ -130,22 +130,22 @@ button{ font-size: 100%;}
     width: 70px;
 }
 
-.circle{
+.card-circle circle{
     fill: none;
     stroke-width: 10;
     /* stroke: #000; */
     transform: translate(5px,5px);
     stroke-dasharray: 190;
-    /* stroke-dashoffset: 190; */
+    stroke-dashoffset: 190;
 }
-/* .card-circle circle:nth-child(1){
+.card-circle circle:nth-child(1){
     stroke-dashoffset: 0;
     stroke: #f3f3f3;
 }
 .card-circle circle:nth-child(2){
-    stroke-dashoffset: calc(190 - ((190*80)/100));
+    /* stroke-dashoffset: calc(190 - ((190*80)/100)); */
     stroke: #49d8a0;
-} */
+}
 
 .active-card .card-circle circle:nth-child(2){
     /* stroke-dashoffset: calc(190 - ((190*80)/100)); */
