--- resources/json/db.json
@@ -1,42 +1,0 @@
-{
-  "projects": [
-    {
-      "id": 1,
-      "project_name": "Sample project 1",
-      "project_desc": "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Duis luctus lorem ipsum, eu suscipit nunc faucibus non. Integer auctor arcu sed sem bibendum, eget ultrices nulla pharetra. Donec rutrum lacus sapien, ut ultrices sem sodales sed. Donec porttitor vel odio vel commodo. Etiam porta porttitor nisi ut tristique. Proin venenatis pharetra tempus.",
-      "tech_used": [
-        "html",
-        "css",
-        "js"
-      ],
-      "percentage_complete": 72,
-      "start_date": "2020-10-03",
-      "end_date": "2020-12-01"
-    },
-    {
-      "id": 2,
-      "project_name": "Sample project 2",
-      "project_desc": "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Duis luctus lorem ipsum, eu suscipit nunc faucibus non. Integer auctor arcu sed sem bibendum, eget ultrices nulla pharetra. Donec rutrum lacus sapien, ut ultrices sem sodales sed. Donec porttitor vel odio vel commodo. Etiam porta porttitor nisi ut tristique. Proin venenatis pharetra tempus.",
-      "tech_used": ["html","css","js"],
-      "percentage_complete": 54,
-      "start_date": "2020-10-03",
-      "end_date": "2020-12-01"
-    }
-  ],
-  "resources":[
-      {
-        "id": 1,
-        "name":"Blachin Skilbadu",
-        "email":"skilbadublachins@qburst.com",
-        "billable":true,
-        "rate_per_hour":200
-    },
-    {
-        "id": 2,
-        "name":"Boogieman",
-        "email":"boogiemans@qburst.com",
-        "billable":true,
-        "rate_per_hour":100
-    }
-    ]
-}
